Remove Serialize module in favor of SerializationStore and in-component serialization.
